third_party/WebKit/LayoutTests/http/tests/devtools/profiler/cpu-profiler-profiling-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that CPU profiling works. https://bugs.webkit.org/show_bug.cgi?id=52634

Profiler was enabled.

Running: testProfiling
found pageFunction cpu-profiler-profiling.js:16
found (anonymous) cpu-profiler-profiling.js:1

Profiler was disabled.
